+++ components/base/PrimaryButton.vue
@@ -1,5 +1,5 @@
 <template>
-  <button class="flex p-4 gap-5 bg-black text-white dark:bg-white dark:text-black font-bold hover:bg-indigo-600 dark:hover:text-white transition-all" @click="handleClick">
+  <button class="font-bebas text-xl flex p-4 gap-5 bg-black text-white dark:bg-white dark:text-black hover:bg-indigo-600 dark:hover:text-white transition-all" @click="handleClick">
     {{ buttonName }}
   </button>
 </template>