+++ components/base/TertiaryButton.vue
@@ -1,28 +1,32 @@
 <template>
-  <button class="flex p-4 gap-5 rounded-full bg-black text-white dark:bg-white dark:text-black font-bold hover:bg-indigo-600 dark:hover:text-white transition-all" @click="handleClick">
-    <component :is="icon" class="w-6 h-6" />
-    {{ buttonName }}
+  <button
+    class="flex p-3 gap-5 rounded-full text-dark hover:text-white bg-gray-300 dark:bg-white dark:text-black font-bold hover:bg-indigo-600 dark:hover:text-white transition-all"
+    @click="handleClick"
+  >
+    <Icon :name="icon" class="text-3xl" />
   </button>
 </template>
 
 <script>
 export default {
   props: {
-    buttonName: {
-      type: String,
-      default: "SECONDARY BUTTON"
-    },
     icon: {
-      type: Object
+      type: String,
+      default: "uil:instagram"
+    },
+    link: {
+      type: String,
+      default: "" 
     }
   },
   methods: {
     handleClick() {
-      if(this.onClick) {
+      if (this.link) {
+        window.open(this.link, "_blank")
+      } else if (this.onClick) {
         this.onClick()
       }
     }
   }
 }
 </script>
-