+++ pages/index.vue
@@ -2,13 +2,18 @@
 
 <template>
   <div>
+    <VideoPlay />
     <BasePrimaryButton />
     <br>
     <BaseSecondaryButton />
     <br>
-    <BaseTertiaryButton />
-    <br>
-    <BaseNavigationButton buttonName="haii" :onClick="() => navigateTo('/certificate')" path="/" />
+    <div class="flex gap-3 items-center">
+      <p class="font-inter font-bold text-white">Follow me here -></p>
+      <BaseTertiaryButton icon="uil:instagram" link="https://instagram.com" />
+      <BaseTertiaryButton icon="uil:github" />
+      <BaseTertiaryButton icon="uil:linkedin" />
+      <BaseTertiaryButton icon="uil:twitter" />
+    </div>
   </div>
 </template>