+++ tailwind.config.js
@@ -1,0 +1,15 @@
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: [],
+  theme: {
+    extend: {
+      fontFamily: {
+        bebas: ["Bebas Neue", "cursive"],
+        inter: ["Inter", "cursive"],
+        jaro: ["Jaro", "cursive"],
+      },
+    },
+  },
+  plugins: [],
+};
+