+++ CHANGELOG.md
@@ -4,4 +4,9 @@
 - Added Header and Footer
 - Switch between light and dark mode
 
+### Feat V1.0.2
+- Added Infinite Video 
+- Added google/fonts for Inter, Jaro, and Bebas Neue
+- Fixed typo in nuxt.config.js
+